requirements.txt
aniso8601==3.0.2
backports.shutil-get-terminal-size==1.0.0
decorator==4.3.0
Django==1.11
djangorestframework==3.8.2
enum34==1.1.6
graphene==2.1.3
graphene-django==2.2.0
graphql-core==2.1
graphql-relay==0.4.5
httplib2==0.11.3
ipython==5.8.0
ipython-genutils==0.2.0
MySQL-python==1.2.5
oauth==1.0.1
oauth2==1.5.211
oauth2client==1.4.12
oauthlib==2.0.0
pathlib2==2.3.2
pexpect==4.6.0
pickleshare==0.7.4
promise==2.2.1
prompt-toolkit==1.0.15
ptyprocess==0.6.0
pyasn1==0.4.4
pyasn1-modules==0.2.2
Pygments==2.2.0
PyJWT==1.6.4
python-dateutil==2.8.0
python-openid==2.2.5
python-social-auth==0.3.6
pytz==2018.5
requests==2.11.1
requests-oauth2==0.2.0
requests-oauthlib==0.5.0
rsa==3.4.2
Rx==1.6.1
scandir==1.9.0
simplegeneric==0.8.1
singledispatch==3.4.0.3
six==1.11.0
social-auth-app-django==2.1.0
social-auth-core==1.7.0
traitlets==4.3.2
typing==3.6.6
virtualenv==16.0.0
wcwidth==0.1.7
